Refactor token retrieval and cache
