+++ src/core/api/services/base/base-service.ts
@@ -1,5 +1,6 @@
 import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import { sessionEventEmitter } from '@/lib/session-events';
+import { TokenCache } from '@/lib/token-cache';
 
 export interface BaseServiceConfig {
   baseURL: string;
@@ -14,57 +15,6 @@ export interface RequestData {
 }
 
 // Token cache to avoid redundant session calls
-class TokenCache {
-  private token: string | null = null;
-  private expiry: number = 0;
-  private refreshPromise: Promise<string | null> | null = null;
-
-  async getToken(refreshFn: () => Promise<string | null>): Promise<string | null> {
-    const now = Date.now();
-    
-    // If token is still valid, return it
-    if (this.token && now < this.expiry) {
-      return this.token;
-    }
-
-    // If there's already a refresh in progress, wait for it
-    if (this.refreshPromise) {
-      return this.refreshPromise;
-    }
-
-    // Start a new refresh
-    this.refreshPromise = this.refreshToken(refreshFn);
-    const newToken = await this.refreshPromise;
-    this.refreshPromise = null;
-    
-    return newToken;
-  }
-
-  private async refreshToken(refreshFn: () => Promise<string | null>): Promise<string | null> {
-    try {
-      const newToken = await refreshFn();
-      
-      if (newToken) {
-        this.token = newToken;
-        // Cache for 5 minutes (considering most JWTs expire in 15-30 minutes)
-        this.expiry = Date.now() + (5 * 60 * 1000);
-      }
-      
-      return newToken;
-    } catch (error) {
-      console.error('Error refreshing token:', error);
-      this.token = null;
-      this.expiry = 0;
-      return null;
-    }
-  }
-
-  invalidate() {
-    this.token = null;
-    this.expiry = 0;
-    this.refreshPromise = null;
-  }
-}
 
 export class BaseService {
   protected instance: AxiosInstance;
@@ -132,28 +82,8 @@ export class BaseService {
     try {
       switch (this.config.authType) {
         case 'bearer':
-          if (typeof window !== 'undefined') {
-            // Client-side: Get token from session
-            const response = await fetch('/api/auth/session');
-            if (response.ok) {
-              const session = await response.json();
-              if (session?.access_token) {
-                return session.access_token;
-              }
-            }
-            
-            // Fallback: Try localStorage/sessionStorage
-            const { tokenStorage } = await import('@/lib/token-storage');
-            return tokenStorage.getToken() || tokenStorage.getTokenSession();
-          } else {
-            // Server-side: Get token from cached session
-            const { getAccessToken } = await import('@/lib/dal');
-            const token = await getAccessToken();
-            if (token) {
-              return token;
-            }
-          }
-          return null;
+          const { fetchAccessToken } = await import('@/lib/auth-token');
+          return await fetchAccessToken();
         case 'api-key':
           return process.env[this.config.authTokenKey || ''] || null;
         case 'client-credentials':
